Move PlotSurface out to own file
